chore: WDS components in Table header (#32112)
Adds WDS components in Table Header
