+++ app/client/src/widgets/wds/WDSTableWidget/component/TableHeader/Pagination.tsx
@@ -1,21 +1,14 @@
-import { TextInput, Text, Flex, IconButton } from "@design-system/widgets";
+import { IconButton, Text } from "@design-system/widgets";
+import { EventType } from "constants/AppsmithActionConstants/ActionConstants";
 import React from "react";
-import { TableFilters } from "./filter";
+import { PageNumberInput } from "./PageNumberInput";
 import type {
   ReactTableColumnProps,
-  TableSizes,
   ReactTableFilter,
-} from "../../Constants";
-import TableDataDownload from "./Download";
-import { EventType } from "constants/AppsmithActionConstants/ActionConstants";
-import { PageNumberInput } from "./PageNumberInput";
-import { Icon as BIcon } from "@blueprintjs/core";
+  TableSizes,
+} from "../Constants";
 
-import { Button, Tooltip } from "@design-system/widgets";
-
-const MIN_WIDTH_TO_SHOW_PAGE_ITEMS = 676;
-
-export interface ActionsPropsType {
+export interface PaginationProps {
   updatePageNo: (pageNo: number, event?: EventType) => void;
   nextPageClick: () => void;
   prevPageClick: () => void;
@@ -47,81 +40,28 @@ export interface ActionsPropsType {
   width: number;
 }
 
-export const Actions = (props: ActionsPropsType) => {
+export const Pagination = (props: PaginationProps) => {
   const pageCount = `${props.pageNo + 1}${
     props.totalRecordsCount ? ` of ${props.pageCount}` : ``
   }`;
 
   return (
     <>
-      {props.isVisibleSearch && (
-        <TextInput
-          onChange={props.searchTableData}
-          placeholder="Search..."
-          startIcon={<BIcon icon="search" />}
-          value={props.searchKey}
-        />
-      )}
-      {(props.isVisibleFilters ||
-        props.isVisibleDownload ||
-        props.allowAddNewRow) &&
-        !!props.columns.length && (
-          <>
-            {props.isVisibleFilters && (
-              <TableFilters
-                applyFilter={props.applyFilter}
-                columns={props.columns}
-                filters={props.filters}
-                widgetId={props.widgetId}
-              />
-            )}
-
-            {props.isVisibleDownload && (
-              <TableDataDownload
-                columns={props.tableColumns}
-                data={props.tableData}
-                delimiter={props.delimiter}
-                widgetName={props.widgetName}
-              />
-            )}
-
-            {props.allowAddNewRow && (
-              <Tooltip
-                tooltip={
-                  props.disableAddNewRow
-                    ? "Save or discard the unsaved row to add a new row"
-                    : ""
-                }
-              >
-                <Button
-                  data-testid="t--add-new-row"
-                  icon="plus"
-                  isDisabled={props.disableAddNewRow}
-                  onPress={props.onAddNewRow}
-                  variant="ghost"
-                >
-                  Add new row
-                </Button>
-              </Tooltip>
-            )}
-          </>
-        )}
-
       {!!props.columns.length &&
         props.isVisiblePagination &&
         props.serverSidePaginationEnabled && (
-          <Flex alignItems="center" gap="spacing-1" marginLeft="auto">
-            {props.totalRecordsCount &&
-              props.width > MIN_WIDTH_TO_SHOW_PAGE_ITEMS && (
-                <Text lineClamp={1} variant="footnote">
-                  {props.totalRecordsCount} Records
-                </Text>
-              )}
+          <div data-table-header-pagination="">
+            {props.totalRecordsCount && (
+              <Text lineClamp={1} variant="footnote">
+                {props.totalRecordsCount} Records
+              </Text>
+            )}
             <IconButton
               icon="chevron-left"
               isDisabled={props.pageNo === 0}
               onPress={props.prevPageClick}
               size="small"
+              variant="outlined"
             />
             <Text lineClamp={1} variant="footnote">
               Page {pageCount}
@@ -134,18 +74,17 @@ export const Actions = (props: ActionsPropsType) => {
               }
               onPress={props.nextPageClick}
               size="small"
+              variant="outlined"
             />
-          </Flex>
+          </div>
         )}
       {!!props.columns.length &&
         props.isVisiblePagination &&
         !props.serverSidePaginationEnabled && (
-          <Flex alignItems="center" gap="spacing-1" marginLeft="auto">
-            {props.width > MIN_WIDTH_TO_SHOW_PAGE_ITEMS && (
-              <Text lineClamp={1} variant="footnote">
-                {props.tableData?.length} Records
-              </Text>
-            )}
+          <div data-table-header-pagination="">
+            <Text lineClamp={1} variant="footnote">
+              {props.tableData?.length} Records
+            </Text>
             <IconButton
               icon="chevron-left"
               isDisabled={props.currentPageIndex === 0}
@@ -156,6 +95,7 @@ export const Actions = (props: ActionsPropsType) => {
                   props.updatePageNo(pageNo + 1, EventType.ON_PREV_PAGE);
               }}
               size="small"
+              variant="outlined"
             />
             <Text lineClamp={1} variant="footnote">
               Page
@@ -181,8 +121,9 @@ export const Actions = (props: ActionsPropsType) => {
                   props.updatePageNo(pageNo + 1, EventType.ON_NEXT_PAGE);
               }}
               size="small"
+              variant="outlined"
             />
-          </Flex>
+          </div>
         )}
     </>
   );
