--- app/client/src/widgets/wds/WDSTableWidget/component/header/actions/Utilities.ts
@@ -1,40 +1,0 @@
-import type { TableColumnProps } from "../../Constants";
-import { isString } from "lodash";
-
-export const transformTableDataIntoCsv = (props: {
-  columns: TableColumnProps[];
-  data: Array<Record<string, unknown>>;
-}) => {
-  const csvData = [];
-  csvData.push(
-    props.columns
-      .map((column: TableColumnProps) => {
-        if (column.metaProperties && !column.metaProperties.isHidden) {
-          return column.Header;
-        }
-        return null;
-      })
-      .filter((i) => !!i),
-  );
-  for (let row = 0; row < props.data.length; row++) {
-    const data: { [key: string]: any } = props.data[row];
-    const csvDataRow = [];
-    for (let colIndex = 0; colIndex < props.columns.length; colIndex++) {
-      const column = props.columns[colIndex];
-      let value = data[column.alias];
-      if (column.metaProperties && !column.metaProperties.isHidden) {
-        value =
-          isString(value) && value.includes("\n")
-            ? value.replace("\n", " ")
-            : value;
-        if (isString(value) && value.includes(",")) {
-          csvDataRow.push(`"${value}"`);
-        } else {
-          csvDataRow.push(value);
-        }
-      }
-    }
-    csvData.push(csvDataRow);
-  }
-  return csvData;
-};
