--- app/client/src/widgets/wds/WDSTableWidget/component/header/banner/styles.module.css
@@ -1,3 +1,0 @@
-.bannerCaption {
-  flex-grow: 1;
-}
